Remove Assignment Instructions & Start on Struct IR
Assignment Instructions have been removed by making the index of the
instruction in the function IR the number of the register storing that
value.

If a function takes two arguments and we want to add them together,
these would be the registers & IR:

f x y = x + y
0 // x
1 // y
2 // FunctionCall "+", 0, 1
3 // Return 2

This is easier to generate since we don't have to mess with assigning
expressions that don't have respective local variables to registers.

Struct IR has also been started. Field getting & setting should be
generated.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>