Add IR Optimizing with HeapToStack Pass
Instructions were changed to be one of two types, Void and Value.
Value instructions return a value that can be used.
Void instructions act on the values returned.

An optimization framework was created, with a HeapToStack Pass
demonstrating how it can be used.

HTS goes through the instructions in a function and keeps track of
which allocations *actually* need to be allocated on the heap. The ones
that don't are converted to stack allocations.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>